minor refactor
