+++ android/app/src/androidTest/java/com/bc/libwally/PsbtTest.kt
@@ -5,7 +5,6 @@ import com.bc.libwally.address.Key
 import com.bc.libwally.address.PubKey
 import com.bc.libwally.bip32.Bip32Path
 import com.bc.libwally.bip32.HDKey
-import com.bc.libwally.bip32.Network
 import com.bc.libwally.core.Core
 import com.bc.libwally.core.Core.base642Bytes
 import com.bc.libwally.core.Core.hex2Bytes
