+++ android/app/src/androidTest/java/com/bc/libwally/TransactionTest.kt
@@ -2,7 +2,6 @@ package com.bc.libwally
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.bc.libwally.address.PubKey
-import com.bc.libwally.bip32.Network
 import com.bc.libwally.core.Core
 import com.bc.libwally.script.ScriptPubKey
 import com.bc.libwally.script.ScriptSig
