+++ java/src/main/java/com/bc/libwally/WallyConstant.java
@@ -5,5 +5,5 @@ public class WallyConstant {
     public static final int WALLY_OK = 0;
     public static final int WALLY_ERROR = -1;
     public static final int WALLY_EINVAL = -2;
-    public static final int WALLY_ENOMEM = -2;
+    public static final int WALLY_ENOMEM = -3;
 }
