+++ java/src/main/java/com/bc/libwally/tx/TxOutput.java
@@ -1,7 +1,7 @@
 package com.bc.libwally.tx;
 
 import com.bc.libwally.address.Address;
-import com.bc.libwally.bip32.Network;
+import com.bc.libwally.Network;
 import com.bc.libwally.script.ScriptPubKey;
 import com.bc.libwally.tx.raw.WallyTxOutput;
