+++ java/src/test/java/com/bc/libwally/AddressTest.java
@@ -3,7 +3,6 @@ package com.bc.libwally;
 import com.bc.libwally.address.Address;
 import com.bc.libwally.address.Key;
 import com.bc.libwally.bip32.HDKey;
-import com.bc.libwally.bip32.Network;
 import com.bc.libwally.script.ScriptPubKey;
 
 import org.junit.Test;
