+++ java/src/test/java/com/bc/libwally/TransactionInstanceTest.java
@@ -2,7 +2,6 @@ package com.bc.libwally;
 
 import com.bc.libwally.address.PubKey;
 import com.bc.libwally.bip32.HDKey;
-import com.bc.libwally.bip32.Network;
 import com.bc.libwally.script.ScriptPubKey;
 import com.bc.libwally.script.ScriptSig;
 import com.bc.libwally.script.ScriptSigType;
