+++ java/src/test/java/com/bc/libwally/TransactionTest.java
@@ -1,7 +1,6 @@
 package com.bc.libwally;
 
 import com.bc.libwally.address.PubKey;
-import com.bc.libwally.bip32.Network;
 import com.bc.libwally.script.ScriptPubKey;
 import com.bc.libwally.script.ScriptSig;
 import com.bc.libwally.script.ScriptSigType;
